moved http library into event.py
